sweep out dead code, vestigial indirection, and clippy debt
Architecture:
- config: drop the never-read lobby_endpoint field (+ its default helpers)
- discord-ipc: drop the never-constructed Emoji/Button activity types
- audio: replace the single-impl Backend trait with an inherent method
- loadout: fold the pass-through update() wrapper into update() itself
  (the other update/update_inner splits do post-dispatch work; this one
  didn't) and derive Default
- tango-pvp: drop Replay::{local,remote}_sram_dump() — callers now
  borrow the fields directly, which also removes two hidden clones on
  the shadow/export paths
- netplay: unbox on_remote_settings' argument at the dispatch site
- session: box ActiveSession::PvP (2.5 KB variant, 10x the others)

Docs that had drifted from the code: config.rs's broken module header,
session::State's doc stranded on MetricSample, folder.rs's orphaned
chip_tooltip_style doc reunited with its fn, refresh_loaded's stale
play.* field names, ActiveSession's missing PvP variant, and a
netplay-sits-atop-net layering note.

Clippy: workspace is now warning-free. getgud/libdatachannel/
tango-filesync gain [lints] workspace = true, mutable_key_type is
allowed workspace-wide (false positive on by-address GameRef keys),
tungstenite's 136-byte error variant is boxed, mgba's unsafe fns get
Safety sections, and the rest is cargo clippy --fix plus hand fixes.

+++ mgba/src/arm_core.rs
@@ -15,11 +15,11 @@ impl<'a> ARMCoreRef<'a> {
     }
 
     pub fn thumb_pc(&self) -> u32 {
-        self.gpr(15) as u32 - mgba_sys::WordSize_WORD_SIZE_THUMB as u32
+        self.gpr(15) as u32 - mgba_sys::WordSize_WORD_SIZE_THUMB
     }
 
     pub fn arm_pc(&self) -> u32 {
-        self.gpr(15) as u32 - mgba_sys::WordSize_WORD_SIZE_ARM as u32
+        self.gpr(15) as u32 - mgba_sys::WordSize_WORD_SIZE_ARM
     }
 
     pub fn execution_mode(&self) -> ExecutionMode {
@@ -53,6 +53,12 @@ impl<'a> ARMCoreMutRef<'a> {
         }
     }
 
+    /// # Safety
+    ///
+    /// The underlying `ARMCore`'s `components` array must be allocated to
+    /// `CPU_COMPONENT_MAX` entries (mgba does this when the core is
+    /// initialized) and must not be freed or reallocated while the
+    /// returned slice is alive.
     pub unsafe fn components_mut(&self) -> &[*mut mgba_sys::mCPUComponent] {
         std::slice::from_raw_parts_mut(
             (*self.ptr).components,
@@ -86,7 +92,7 @@ impl<'a> ARMCoreMutRef<'a> {
                 as *const u16) as u32;
 
             // pc += WORD_SIZE_THUMB;
-            pc += mgba_sys::WordSize_WORD_SIZE_THUMB as u32;
+            pc += mgba_sys::WordSize_WORD_SIZE_THUMB;
 
             // LOAD_16(cpu->prefetch[1], pc & cpu->memory.activeMask, cpu->memory.activeRegion);
             (*self.ptr).prefetch[1] = *(((*self.ptr).memory.activeRegion as *const u8)
